home/backend/private/services/scheduler/requirements.txt
aiocron
aiosqlite
msgspec
orjson
uvloop
valkey[libvalkey]
websockets
netifaces2
aiofiles
